Judge frame noise on the symbol without its template arguments
A closure type passed as a template argument spells its own
`::{lambda()#1}` inside the argument list, so `Bar<setup()::{lambda()#1}>
::run` read as a trampoline and the title named its caller instead.
Anchoring the pattern doesn't separate the two — the `::` is there in
both — so the test now runs against the symbol with template arguments
stripped, which is what tells a real instantiation from the `_FUN` thunk.

Say Device rather than the bug form's `Other` catch-all. `Other` earns
its place beside the form's platform field and says nothing as a title
prefix; nRF52 targets land there today.

Move the crash handler's own annotations — the reason line, and the
`(backtrace)` / `(stack scan)` frame labels — to crash-detector, which
already owns how crash output is read, and take excerpt lines rather
than a scrape so nothing points back at report assembly. That returns
crash-report.ts to 578 lines. Lift the dialog's styles into
crash-report-dialog.styles.ts, following its six siblings, and split the
form off its render; the file was 603 lines with a 122-line render
against a 500-600 cap and a ~100-line render guideline.

Hoist the regexes those functions compiled inline, matching how every
other pattern in both files is declared.
