+++ src/util/crash-detector.ts
@@ -66,10 +66,14 @@ export const DECODED_RE = /^(?:WARNING )?Decoded (0x[0-9a-fA-F]{8}.*)$/;
 // because gcc appends ` (discriminator N)` past the line number.
 const DECODED_FRAME_RE = /^0x[0-9a-fA-F]+:\s*(.+?)(?:\s+at\s+.*)?$/;
 
+// The same frame's address alone, for matching one against the `BT<n>:`
+// line that labels it.
+const DECODED_FRAME_ADDRESS_RE = /^0x([0-9a-fA-F]+):/;
+
 /** The address a decoded frame carries, lowercased and without `0x`; ""
  *  when the entry isn't a decoded frame. */
 export function decodedFrameAddress(frame: string): string {
-  const match = /^0x([0-9a-fA-F]+):/.exec(frame.split("\n")[0].trim());
+  const match = DECODED_FRAME_ADDRESS_RE.exec(frame.split("\n")[0].trim());
   return match ? match[1].toLowerCase() : "";
 }
 
@@ -104,4 +108,51 @@ export function latchCrashKind(
   next: CrashKind | null
 ): CrashKind | null {
   return current === "live" || next === null ? current : next;
+}
+
+// The esp32 crash handler labels every stored frame: `(backtrace)` for one
+// the unwinder produced, `(stack scan)` for a word that merely looks like a
+// return address in stack memory. A scanned word decodes to whatever symbol
+// owns that address, so it can name a component the crash never entered.
+const STACK_SCAN_RE = /BT\d+:\s*(?:0x)?([0-9a-fA-F]{8})\b.*\(stack scan\)/;
+
+// The handler's own verdict, `Reason: <type>` or `<type> - <detail>`
+// (`Task wdt`, `Fault - Store access fault`). Two crashes can share a frame
+// and differ entirely in why they got there.
+const CRASH_REASON_RE = /crash:\d+\]:\s*Reason:\s*(.+?)\s*$/;
+
+// A watchdog exception class, and the raw cause the handler appends to a
+// reason: "(exccause=4)", "(cause 0)".
+const WATCHDOG_TYPE_RE = /wdt/i;
+const TRAILING_CAUSE_RE = /\s*\([^)]*\)\s*$/;
+
+/** What the crash handler blamed, preferring its specific cause over the
+ *  exception class; "" when the dump carries no reason. */
+export function crashReason(excerpt: string[]): string {
+  for (const line of excerpt) {
+    const match = CRASH_REASON_RE.exec(line);
+    if (!match) continue;
+    const [type, detail] = match[1].split(" - ");
+    // A watchdog names the condition itself; its detail is only the trap
+    // that fired ("Soft WDT - Level1Int"), so there the type is what reads.
+    const chosen = WATCHDOG_TYPE_RE.test(type) ? type : (detail ?? type);
+    const reason = chosen.replace(TRAILING_CAUSE_RE, "").trim();
+    // "Unknown" is the handler saying it couldn't decode the cause.
+    return reason === "Unknown" ? "" : reason;
+  }
+  return "";
+}
+
+/**
+ * The decoded frames the unwinder vouched for. All of them when the dump
+ * draws no distinction — live panics and esp8266 print no such label.
+ */
+export function unwoundFrames(excerpt: string[], decodedFrames: string[]): string[] {
+  const scanned = new Set<string>();
+  for (const line of excerpt) {
+    const match = STACK_SCAN_RE.exec(line);
+    if (match) scanned.add(match[1].toLowerCase());
+  }
+  if (scanned.size === 0) return decodedFrames;
+  return decodedFrames.filter((frame) => !scanned.has(decodedFrameAddress(frame)));
 }
